+++ src/hw12/Baraban.java
@@ -1,0 +1,13 @@
+package hw12;
+
+class Baraban implements Musicalinstrument {
+    private final String size;
+
+    public Baraban(String size) {
+        this.size = size;
+    }
+
+    public void play() {
+        System.out.println("Играет барабан размером " + size);
+    }
+}
