+++ src/hw12/Book.java
@@ -1,0 +1,14 @@
+package hw12;
+
+public class Book implements Printable {
+    private final String title;
+
+    public Book(String title) {
+        this.title = title;
+    }
+
+    public void print() {
+        System.out.println("Book: " + title);
+    }
+
+}
