+++ src/hw12/Gitara.java
@@ -1,0 +1,13 @@
+package hw12;
+
+class Gitara implements  Musicalinstrument{
+    private final int numberOfStrings ;
+
+    public Gitara(int numberOfStrings) {
+        this.numberOfStrings = numberOfStrings;
+    }
+
+    public void play() {
+        System.out.println("Играет гитара с " + numberOfStrings + " струнами");
+    }
+}
